requirements.txt
# pip install -r requirements.txt

pika==1.2.0
python-binance==1.0.12
psycopg2==2.9.1
python-dotenv==0.19.0

google-api-python-client==2.17.0
oauth2client==4.1.3
google-auth-oauthlib==0.4.5

APScheduler==3.7.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-APScheduler==1.12.2


